Set return types to float
This is because it is unclear whether a question score can be a float or not.
With floats we capture at least all cases and it makes this assumption clear in the code.

+++ app/Http/Helpers/Normalize.php
@@ -78,7 +78,7 @@ class Normalize
             || is_null($this->testTake->getAttribute($value));
     }
 
-    private function getTotalScore(): ?int
+    private function getTotalScore(): ?float
     {
         if (
             ($this->isNotInRequest('ppp') && $this->isInRequest('epp'))
